Updated Returns Code File
Realized that my interpretation of the results were wrong. Didn't change much of the results though lol. I'm thinking we can nullify the intercept though when I calculate via the LR equation that was derived. It gives me 0.07% which is A LOT closer to the final result that we are looking for in comparison to actual price.